rename shade slider to shade picker and fix tests
