+++ src/app/app.component.html
@@ -186,11 +186,11 @@
     <div class="col-12 col-md-7 mb-3">
       <div class="d-flex justify-content-center justify-content-md-start">
         <div style="width: 320px;display: inline-block">
-          <color-shade-slider
+          <color-shade-picker
             [color]="state"
             (onChangeComplete)="changeComplete($event)"
             (onChange)="changeComplete($event)"
-          ></color-shade-slider>
+          ></color-shade-picker>
           <div class="pt-3 text-center label">Shade Slider</div>
         </div>
       </div>
