--- src/lib/components/shade-slider/shade-slider.spec.ts
@@ -1,71 +1,0 @@
-import { Component } from '@angular/core';
-import { async, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-
-import { ColorShadeSliderModule, ShadeSliderComponent } from './shade-slider.component';
-
-export const red = {
-  hsl: { a: 1, h: 0, l: 0.5, s: 1 },
-  hex: '#ff0000',
-  rgb: { r: 255, g: 0, b: 0, a: 1 },
-  hsv: { h: 0, s: 1, v: 1, a: 1 },
-};
-
-
-describe('AlphaComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ColorShadeSliderApp],
-      imports: [ColorShadeSliderModule],
-    });
-
-    TestBed.compileComponents();
-  }));
-  it(`should apply className to root element`, () => {
-    const fixture = TestBed.createComponent(ShadeSliderComponent);
-    const testComponent = fixture.componentInstance;
-    testComponent.className = 'classy';
-    fixture.detectChanges();
-    const div = fixture.debugElement.query(By.css('.alpha-picker'));
-    expect(div.nativeElement.classList.contains('classy')).toBe(true);
-  });
-  it(`should draw vertical`, () => {
-    const fixture = TestBed.createComponent(ColorShadeSliderApp);
-    const testComponent = fixture.componentInstance;
-    fixture.detectChanges();
-    fixture.detectChanges();
-    const div = fixture.debugElement.query(By.css('.alpha-container'));
-    expect(div.nativeElement.classList.contains('color-alpha-vertical')).toBe(true);
-  });
-  // it(`should change alpha on mousedown`, () => {
-  //   const fixture = TestBed.createComponent(AlphaPickerComponent);
-  //   const component = fixture.componentInstance;
-  //   component.width = 20;
-  //   component.height = 200;
-  //   component.color = red.hsl;
-  //   fixture.detectChanges();
-  //   const $event = new MouseEvent('mousedown', {
-  //     bubbles: true,
-  //     cancelable: true,
-  //     view: window,
-  //     clientX: 0,
-  //     clientY: 0,
-  //   });
-  //   fixture.detectChanges();
-  //   expect(component.hsl.a).toEqual(0);
-  // });
-});
-
-
-@Component({
-  selector: 'test-app',
-  template: `
-  <color-shade-slider
-    [className]="className"
-  >
-  </color-shade-slider>
-  `,
-})
-class ColorShadeSliderApp {
-  className = '';
-}
